refactor: modularize game and fix environment transition
