+++ assets/sfx_hurt_synth.wav
