+++ ecco/__init__.py
@@ -1,0 +1,3 @@
+from .game import run
+
+__all__ = ["run"]
